Rename datafile scraper base_parser module to base
